update arg variable

+++ src/map/mob.cpp
@@ -3581,13 +3581,13 @@ int mob_skill_id2skill_idx(int mob_id,uint16 skill_id)
  *------------------------------------------*/
 int mob_getfriendhprate_sub(struct block_list *bl,va_list ap)
 {
-	int min_rate, max_rate,rate;
+	int64 min_rate, max_rate,rate;
 	struct block_list **fr;
 	struct mob_data *md;
 
 	md = va_arg(ap,struct mob_data *);
-	min_rate=va_arg(ap,int);
-	max_rate=va_arg(ap,int);
+	min_rate=va_arg(ap,int64);
+	max_rate=va_arg(ap,int64);
 	fr=va_arg(ap,struct block_list **);
 
 	if( md->bl.id == bl->id && !(battle_config.mob_ai&0x10))
@@ -3637,7 +3637,7 @@ struct block_list *mob_getmasterhpltmaxrate(struct mob_data *md,int64 rate)
  *------------------------------------------*/
 int mob_getfriendstatus_sub(struct block_list *bl,va_list ap)
 {
-	int cond1,cond2;
+	int64 cond1,cond2;
 	struct mob_data **fr, *md, *mmd;
 	int flag=0;
 
@@ -3650,8 +3650,8 @@ int mob_getfriendstatus_sub(struct block_list *bl,va_list ap)
 
 	if (battle_check_target(&mmd->bl,bl,BCT_ENEMY)>0)
 		return 0;
-	cond1=va_arg(ap,int);
-	cond2=va_arg(ap,int);
+	cond1=va_arg(ap,int64);
+	cond2=va_arg(ap,int64);
 	fr=va_arg(ap,struct mob_data **);
 	if( cond2==-1 ){
 		int j;
